+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add function `ndmtag.clear` that clears the tag from the DOM with the given name.
+- Add function `ndmtag.refresh` that renders a new ad for the tag with the given name. 
 
 ## [2.1.0] - 2019-11-21
 ## [2.1.0-beta.1] - 2019-10-16